sketch_server/requirements.txt
efficientnet_torch
sklearn
tqdm
torchvision